requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
aiohttp==3.9.5
psycopg2-binary==2.9.9